app/requirements.txt
anthropic>=0.40.0
azure-identity>=1.19.0
azure-mgmt-resource>=23.0.0
azure-mgmt-compute>=32.0.0
azure-mgmt-network>=27.0.0
azure-mgmt-storage>=21.0.0
azure-mgmt-containerservice>=32.0.0
azure-mgmt-authorization>=4.0.0
# azure-mgmt-policyinsights dropped — all versions require 'six' (py2 compat shim).
# Compliance queries use the Policy Insights REST API directly instead.
azure-mgmt-subscription>=3.0.0
python-dotenv>=1.0.0
requests>=2.32.0
